+++ src/txdb.h
@@ -194,8 +194,8 @@ public:
     bool WriteTxIndex(const std::vector<std::pair<uint256, CDiskTxPos> > &list);
     bool WriteFlag(const std::string &name, bool fValue);
     bool ReadFlag(const std::string &name, bool &fValue);
-    bool WriteHashMask(const uint256 &_hashMask);
-    bool ReadHashMask(uint256 &_hashMask);
+    bool WriteHashMask(const uint64_t &_hashMask);
+    bool ReadHashMask(uint64_t &_hashMask);
     bool WriteHashMaskThreshold(const uint8_t &_hashMaskThreshold);
     bool ReadHashMaskThreshold(uint8_t &_hashMaskThreshold);
     bool FindBlockIndex(uint256 blockhash, CDiskBlockIndex *index);
